fixed parabola class

+++ visualizer/include/parabola.h
@@ -10,11 +10,11 @@ namespace models {
         [[maybe_unused]] double secondCoefficient{};
         [[maybe_unused]] double thirdCoefficient{};
         [[nodiscard]] double countParabolaValue(double x) const;
-        void calculateParabolaPoints(const cv::Point2i &firstPoint, const cv::Point2i &secondPoint,
-                                                const cv::Point2i &thirdPoint);
+        void calculateParabolaPoints(const cv::Point2d &firstPoint, const cv::Point2d &secondPoint,
+                                                const cv::Point2d &thirdPoint);
         [[maybe_unused]] std::vector<cv::Point3d> borderPoints;
     public:
-        [[maybe_unused]] parabola(const cv::Point2i& firstPoint, const cv::Point2i& secondPoint, const cv::Point2i& thirdPoint);
+        [[maybe_unused]] parabola(const cv::Point2d& firstPoint, const cv::Point2d& secondPoint, const cv::Point2d& thirdPoint);
         [[maybe_unused]] [[nodiscard]] cv::Vec3d parabolaCoefficient() const;
     };
 }
